smartads/backend/requirements.txt
flask
pymongo
flask-cors
python-dotenv
werkzeug
google-generativeai
cloudinary
bcrypt
